+++ src/main/java/org/truffleruby/core/GCNodes.java
@@ -18,23 +18,9 @@ import org.truffleruby.builtins.CoreModule;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.object.DynamicObject;
 
 @CoreModule("GC")
 public abstract class GCNodes {
-
-    @CoreMethod(names = "start", onSingleton = true)
-    public static abstract class GCStartNode extends CoreMethodArrayArgumentsNode {
-
-        @TruffleBoundary
-        @Specialization
-        protected DynamicObject vmGCStart() {
-            getContext().getMarkingService().queueMarking();
-            System.gc();
-            return nil();
-        }
-
-    }
 
     @CoreMethod(names = "count", onSingleton = true)
     public abstract static class CountNode extends CoreMethodArrayArgumentsNode {